pyproject.toml
[project]
name = "venuemap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "lxml>=6.1.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]

[tool.hatch.build.targets.wheel]
packages = ["src/venuemap"]
